Move functions to utils file
