+++ site/gatsby-site/src/utils/serverless.ts
@@ -1,13 +1,16 @@
+import config from "../../server/config"
+import { HandlerEvent } from '@netlify/functions'
+
 export const createResponse = () => {
 
     let responseHeaders: Record<string, string> = { 'Content-Type': 'application/json' }
     let response: Record<string, any> = {}
 
     const res = {
-        getHeader: (name: any) => {
+        getHeader: (name: string) => {
             return responseHeaders[name]
         },
-        setHeader: (name: any, value: any) => {
+        setHeader: (name: string, value: string) => {
             if (name.toLowerCase() === 'set-cookie') {
                 response.multiValueHeaders = {
                     ...response.multiValueHeaders,
@@ -18,22 +21,15 @@ export const createResponse = () => {
                 response.headers = responseHeaders
             }
         },
-        status: (statusCode: any) => {
+        status: (statusCode: number) => {
             response.statusCode = statusCode
             return res
         },
-        send: (body: any) => {
+        send: (body: string | Record<string, any>) => {
             response.body = typeof body === 'string' ? body : JSON.stringify(body)
         },
-        json: (json: any) => {
+        json: (json: Record<string, any>) => {
             response.body = JSON.stringify(json)
-        },
-        redirect: (statusOrUrl: any, url: any) => {
-            const statusCode = typeof statusOrUrl === 'number' ? statusOrUrl : 302
-            const redirectUrl = typeof statusOrUrl === 'string' ? statusOrUrl : url
-            response.statusCode = statusCode
-            responseHeaders.Location = redirectUrl
-            response.headers = responseHeaders
         },
         end: () => { },
         getResponse: () => response,
@@ -41,4 +37,79 @@ export const createResponse = () => {
     }
 
     return res
+}
+
+const parseBody = (event: HandlerEvent) => {
+
+    const contentType = event.headers['content-type'] || event.headers['Content-Type'];
+    const { body } = event;
+
+    if (!body) return undefined;
+
+    if (contentType?.includes('application/json')) {
+        try {
+            return JSON.parse(body);
+        } catch (e) {
+            console.error('Failed to parse JSON body:', e);
+            return undefined;
+        }
+    }
+
+    if (contentType?.includes('application/x-www-form-urlencoded')) {
+        const params = new URLSearchParams(body);
+        const formData: Record<string, string> = {};
+        params.forEach((value, key) => {
+            formData[key] = value;
+        });
+        return formData;
+    }
+
+    return undefined;
+};
+
+const parseCookies = (cookieHeader: string) => {
+    const cookies: Record<string, string> = {};
+    if (!cookieHeader) return cookies;
+
+    cookieHeader.split(';').forEach(cookie => {
+        const parts = cookie.trim().split('=');
+        if (parts.length >= 2) {
+            const key = parts[0];
+            const value = parts.slice(1).join('=');
+            cookies[key] = decodeURIComponent(value);
+        }
+    });
+
+    return cookies;
+};
+
+const parseHeaders = (event: any) => {
+
+    return {
+        ...event.headers,
+        host: event.headers.host || new URL(config.NEXTAUTH_URL).host,
+    }
+}
+
+const parseQuery = (event: any) => {
+
+    const path = event.path.replace('/api/auth/', '')
+    const nextAuthArray = path.split('/').filter(Boolean);
+
+    return {
+        ...event.queryStringParameters,
+        nextauth: nextAuthArray,
+        action: nextAuthArray[nextAuthArray.length - 1],
+        providerId: nextAuthArray[1]
+    }
+}
+
+export const recreateRequest = (event: HandlerEvent) => {
+    return {
+        method: event.httpMethod,
+        query: parseQuery(event),
+        cookies: parseCookies(event.headers.cookie ?? ''),
+        headers: parseHeaders(event),
+        body: parseBody(event),
+    }
 }
